Rename zsrb to zsrb_simple
The reason for this is that this function only calculates the simple SRB as
described by Duff 2023.

A new function will be added here that is able to calcualate both the simple and
complex SRB (as per Hammers ... Duff, 2016). This simple SRB is derived
differently from the 2023 method described by Duff.