Added a seed for the random permutations. Reverted to model p.
- The seed ensures that even though the permutations are random, you will have
replicable results.
- Reverting to the model p means that we take the summary of the model and get
the p-values from there, instead of running car::Anova and use that p-value. The
downside is that this original p-value is truncted and it does not seem possible
to fix this (see https://github.com/mtorchiano/lmPerm/issues/5). The upside is
that we now at least that these are the true p-values for this method.